Update Capture1
I do not like some of the maps on the current rotation so I decided to do a little change. First, I believed the rotation needed more maps, so I added some. I took out Mushroom Gorge because the void gap is excessively long. If it were shorter, it would be a much better map. Battle Ecliptica III has a nice twist to it, but the lanes are very short, and it is very easy to prevent people from getting to a wool room. Also, you cannot place along the side of the lanes without block glitching. The lanes are also a bit thin. I removed Tenebrous because of the force single-cap. And there is only one way to get into the wool room which is usually filled with about 11 per lane. Also, I spaced out the maps that were in the rotations before. Having the maps from the older rotations play one after another gives the feel of the old rotation and gets boring because of the amount of times the map was played before.