Removing BlockRAGE
Removing BlockRAGE until the map gets an update for the middle spawners.
As the map is rage the middle spawner is useless and makes players
believe the items there are better. I recommend having items such as
resistance potions in there so you can go on a small spree or stuff like
that. Maybe jump boost?
The second reason is there is no sword and players quickly get confused
(after playing it today and people said wheres the sword in chat).

Otherwise the map is good and I quite enjoy it just update it or if the
update will take a long time, remove it till then.