Add RFV3 to EU Capture
While some of the mapdevs might strongly disagree with this, regardless of bridging issues or not, this map was very fun and is a classic to me; I loved the long matches and defenses. I don't think it should be on every server, but I honestly think it should be into the rotations just once. This was my favorite map at one point and it's a shame to see it go.